pyproject.toml
[tool.poetry]
name = "etsy-image-generator"
version = "2.0.0"
package-mode = false

[project]
name = "etsy-image-generator"
version = "2.0.0"
description = "Pro-grade Etsy product image generator with AI-powered prompt engineering and multi-provider support"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "replicate>=0.25.0",
    "openai>=1.10.0",
    "google-genai>=1.0.0",
    "Pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
include = ["core*", "prompts*", "app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
